requirements.txt
pillow
pyyaml
pyenchant
matplotlib
discord
pandas